crimson/osd: move pipelines to osd_operation.h
Each of the two existing pipelines are shared across multiple
ops.  Rather than defining them in a specific op or in
osd_operations/common/pg_pipeline.h, just declare them in
osd_operation.h.
